requirements.txt
wheel
requests
six
pytz
psycopg2>=2.4.5
pillow
django>=1.9.0
sphinx
markdown
reportlab
djangorestframework
django-allauth
django-stdimage
django-autocomplete-light==2.3.3
python3-memcached
django-compressor
django-db-mailer
monthdelta
django-tagging
django-contrib-comments
django-recaptcha
django-bit-category
django-invoice2
django-auditlog
django-jsonfield
django-basic-stats
urljects
django-ckeditor
django-generic-ratings